requirements.txt
faiss-cpu==1.7.4
sentence-transformers==2.2.2
scikit-learn==1.3.2
numpy==1.24.3
torch==2.0.1
transformers==4.34.0
